billing cycle check separate from ad poller
