+++ coldfront/plugins/qumulo/management/commands/add_scheduled_ad_poller.py
@@ -7,10 +7,6 @@ from coldfront.plugins.qumulo.tasks import (
     conditionally_update_storage_allocation_statuses,
 )
 
-from coldfront.plugins.qumulo.management.commands.check_billing_cycles import (
-    check_allocation_billing_cycle_and_prepaid_exp,
-)
-
 
 class Command(BaseCommand):
     def handle(self, *args, **options):
@@ -22,19 +18,8 @@ class Command(BaseCommand):
             minutes=1,
             repeats=-1,
         )
-        print("Scheduling Prepaid Expiration Date Scanner")
-        Schedule.objects.get_or_create(
-            func="coldfront.plugins.qumulo.management.commands.add_scheduled_ad_poller.prepaid_expiration_cleanup",
-            name="Check Billing Statuses",
-            schedule_type=Schedule.DAILY,
-            repeats=-1,
-        )
 
 
 def sequential_poll_and_check() -> None:
     poll_ad_groups()
     conditionally_update_storage_allocation_statuses()
-
-
-def prepaid_expiration_cleanup() -> None:
-    check_allocation_billing_cycle_and_prepaid_exp()
