+++ coldfront/plugins/qumulo/management/commands/run_ad_poller.py
@@ -6,9 +6,6 @@ from coldfront.plugins.qumulo.tasks import (
     poll_ad_groups,
     conditionally_update_storage_allocation_statuses,
 )
-from coldfront.plugins.qumulo.management.commands.check_billing_cycles import (
-    check_allocation_billing_cycle_and_prepaid_exp,
-)
 
 
 class Command(BaseCommand):
@@ -22,6 +19,5 @@ class Command(BaseCommand):
             [
                 (poll_ad_groups),
                 (conditionally_update_storage_allocation_statuses),
-                (check_allocation_billing_cycle_and_prepaid_exp),
             ]
         )
